Control.Monad.Sample

Documentation

class MonadPlus m => MonadSample r m | m -> r where

MonadPlus to imply that an empty pool is empty forever unless you re-fill it first.

Minimal complete definition

sample

Methods

sample :: m r

sampleN :: Vector v r => Int -> m (v r)

Should never fail

flushSamples :: MonadSample r m => m [r]

sampleReservoir :: forall m v r. (PrimMonad m, MonadSample r m, Vector v r) => Int -> Gen (PrimState m) -> m (v r)

Pulls samples until it cannot pull any more, and returns N shuffled items from the pool. Is O(N) memory.

data SampleAct m r

Constructors

SA

Fields

saSample :: m (Maybe r)

saSampleN :: forall v. Vector v r => Int -> m (Maybe (v r))

newtype SampleT r m a

Constructors

ST

Fields

sampleTReader :: MaybeT (ReaderT (SampleAct m r) m) a

Bundled Patterns

pattern SampleT :: (SampleAct m r -> m (Maybe a)) -> SampleT r m a

runSampleT :: SampleT r m a -> SampleAct m r -> m (Maybe a)

simpleSA :: forall m r. Monad m => m (Maybe r) -> SampleAct m r

saRef :: Ref m [r] v => v -> SampleAct m r

saState :: MonadState [r] m => SampleAct m r

saConduit :: Monad m => SampleAct (ConduitT r o m) r