Numeric.Opto

Documentation

module Numeric.Opto.Update

module Numeric.Opto.Run.Conduit

module Numeric.Opto.Run

module Numeric.Opto.Optimizer

module Numeric.Opto.Core

module Control.Monad.Sample

class Monad m => Ref m a v | v -> a where

Abstraction over types of mutable references for values. A Ref m a v means that a v is a mutable reference to an a, and we may updatewritemodify v in context m.

This allows us to reat mutable vectors and in-place mutable numbers or records in the same way.

Minimal complete definition

thawRef, (copyRef | updateRef, updateRef')

Methods

thawRef :: a -> m v

Initialize a mutable reference with a given value

freezeRef :: v -> m a

Read an immutable value back from a mutable reference

copyRef :: v -> a -> m ()

Copy an immutable value into a mutable reference

modifyRef :: v -> (a -> a) -> m ()

Apply a pure function on an immutable value onto a value stored in a mutable reference.

modifyRef' :: v -> (a -> a) -> m ()

modifyRef, but forces the result before storing it back in the reference.

updateRef :: v -> (a -> (a, b)) -> m b

Apply a pure function on an immutable value onto a value stored in a mutable reference, returning a result value from that function.

updateRef' :: v -> (a -> (a, b)) -> m b

updateRef, but forces the updated value before storing it back in the reference.