Numeric.Opto.Optimizer

Description

Defining various numeric optimizers. Most of these implemtations are taken directly from http://ruder.io/optimizing-gradient-descent/

Documentation

steepestDescent

Arguments

:: LinearInPlace m v c a

=> c
learning rate

-> Grad m r a
gradient

-> Opto m v r a

Steepest descent, acording to some learning rate. The simplest optimizer.

newtype Momentum c

Hyperparameter for momentum

Constructors

Momentum

Fields

momentumDecay :: c

momentum

Arguments

:: LinearInPlace m v c a

=> Momentum c
configuration

-> c
learning rate

-> Grad m r a
gradient

-> Opto m v r a

Steepest descent with momentum. (Qian, 1999)

newtype Nesterov c

Hyperparameter for nesterov

Constructors

Nesterov

Fields

nesterovDecay :: c

nesterov

Arguments

:: LinearInPlace m v c a

=> Nesterov c
configuration

-> c
learning rate

-> Grad m r a
gradient

-> Opto m v r a

Nesterov accelerated gradient (NAG) (Nesterov, 1983)

data Adagrad c

Hyperparameters for adagrad

Constructors

Adagrad

Fields

adagradRate :: c

adagradEps :: c

adagrad :: forall m v r a c. (LinearInPlace m v c a, Floating a, Real c) => Adagrad c -> Grad m r a -> Opto m v r a

Adaptive Gradient (Duchu, Hazan, Singer, 2011). Note that if the state is not reset periodically, updates tend to zero fairly quickly.

data Adadelta c

Hyperparameters for adadelta

Constructors

Adadelta

Fields

adadeltaDecay :: c

adadeltaEps :: c

adadelta :: forall m v r a c. (LinearInPlace m v c a, Floating a, Real c) => Adadelta c -> Grad m r a -> Opto m v r a

The Adadelta extension of Adagrad (Zeiler, 2012) that mitigates the decreasing learning rate.

data RMSProp c

Hyperparameters for rmsProp

Constructors

RMSProp

Fields

rmsPropRate :: c

rmsPropDecay :: c

rmsPropEps :: c

rmsProp :: forall m v r a c. (LinearInPlace m v c a, Floating a, Real c) => RMSProp c -> Grad m r a -> Opto m v r a

RMSProp, as described by Geoff Hinton.

data Adam c

Hyperparameters for adam

Constructors

Adam

Fields

adamStep :: !c

adamDecay1 :: !c

adamDecay2 :: !c

adamEps :: !c

adam

Arguments

:: (RealFloat c, Floating a, LinearInPlace m v c a, PrimMonad m)

=> Adam c
configuration

-> Grad m r a
gradient

-> Opto m v r a

Adaptive Moment Estimation (Kingma, Ba, 2015)

data AdaMax c

Hyperparameters for adaMax

Constructors

AdaMax

Fields

adaMaxStep :: !c

adaMaxDecay1 :: !c

adaMaxDecay2 :: !c

adaMaxEps :: !c

adaMax

Arguments

:: (RealFloat c, Metric c a, LinearInPlace m v c a, PrimMonad m)

=> AdaMax c
configuration

-> Grad m r a
gradient

-> Opto m v r a

Adam variation (Kingma and Ba, 2015)