Numeric.Opto.Ref

Documentation

class Monad m => Ref m a v | v -> a where

Abstraction over mutable references of types. A Ref m a v means that a v is a mutable reference to an a, and we may updatewritemodify v in context m.

This allows us to reat mutable vectors and in-place mutable numbers or records in the same way.

Minimal complete definition

thawRef, (copyRef | updateRef, updateRef')

Methods

thawRef :: a -> m v

Initialize a mutable reference with a given value

freezeRef :: v -> m a

Read an immutable value back from a mutable reference

copyRef :: v -> a -> m ()

Copy an immutable value into a mutable reference

modifyRef :: v -> (a -> a) -> m ()

Apply a pure function on an immutable value onto a value stored in a mutable reference.

modifyRef' :: v -> (a -> a) -> m ()

modifyRef, but forces the result before storing it back in the reference.

updateRef :: v -> (a -> (a, b)) -> m b

Apply a pure function on an immutable value onto a value stored in a mutable reference, returning a result value from that function.

updateRef' :: v -> (a -> (a, b)) -> m b

updateRef, but forces the updated value before storing it back in the reference.

data RefVal :: (Type -> Type) -> Type -> Type -> Type where

RefVal m a v is a wrapper over a pure value a that can be stored in a mutable reference v.

Constructors

RVl :: Ref m a v => a -> RefVal m a v

data RefVar :: (Type -> Type) -> Type -> Type -> Type where

RefVar m a v is a wrapper over a mutable reference v that contains a value a.

Constructors

RVr :: Ref m a v => v -> RefVar m a v

type RefVals m = ZipProd (RefVal m)

Store multiple RefVals

type RefVars m = ZipProd (RefVar m)

Store multiple RefVars

thawRefs :: Applicative m => RefVals m as vs -> m (RefVars m as vs)

thawRef, but over a collection RefVals.

freezeRefs :: Applicative m => RefVars m as vs -> m (Rec Identity as)

freezeRef, but over a collection RefVars.

pullRefs :: Applicative m => RefVars m as vs -> m (RefVals m as vs)

A version of freezeRefs that returns RefVals.