Numeric.Opto.Update

Description

A unified interface for values in vector spaces that can be added and scaled (purely and also in-place), and measured.

Documentation

class Num c => Linear c a | a -> c where

If a is an instance of Linear c, you can add together values of a, and scale them using cs.

For example, if you have a vector of doubles, you can add them together component-wise, and scale them by multiplying every item by the scalar.

Mathematically, this means that a forms something like a module or vector space over c, where c can be any Num instance.

Minimal complete definition

Nothing

Methods

(.+.) :: a -> a -> a infixl 6

Add together as. Should be associative.

x .+. (y .+. z) == (x .+. y) .+. z

If a is an instance of Num, this can be just +.

zeroL :: a

The "zero" a, meant to form an identity with .+..

x .+. zeroL == x
zeroL .+. y == y

If a is an instance of Num, this can be just 0.

(.*) :: c -> a -> a infixl 7

Scale an a by a factor c. Should distribute over .+..

a .* (x .+. y) == (a .* x) .+. (a .* y)
a .* (b .* c)  == (a * b) .* c

(.+.) :: (ADTRecord a, Constraints a (Linear c)) => a -> a -> a infixl 6

Add together as. Should be associative.

x .+. (y .+. z) == (x .+. y) .+. z

If a is an instance of Num, this can be just +.

zeroL :: (ADTRecord a, Constraints a (Linear c)) => a

The "zero" a, meant to form an identity with .+..

x .+. zeroL == x
zeroL .+. y == y

If a is an instance of Num, this can be just 0.

(.*) :: (ADTRecord a, Constraints a (Linear c)) => c -> a -> a infixl 7

Scale an a by a factor c. Should distribute over .+..

a .* (x .+. y) == (a .* x) .+. (a .* y)
a .* (b .* c)  == (a * b) .* c

sumLinear :: (Linear c a, Foldable t) => t a -> a

Sum over a Foldable container of Linear c a

gAdd :: forall c a. (ADTRecord a, Constraints a (Linear c)) => a -> a -> a

An implementation of .+. that works for records where every field is an instance of Linear c (that is, every field is additive and can be scaled by the same c).

gZeroL :: forall c a. (ADTRecord a, Constraints a (Linear c)) => a

An implementation of zeroL that works for records where every field is an instance of Linear c (that is, every field is additive and can be scaled by the same c).

gScale :: forall c a. (ADTRecord a, Constraints a (Linear c)) => c -> a -> a

An implementation of .* that works for records where every field is an instance of Linear c (that is, every field is additive and can be scaled by the same c).

class Linear c a => Metric c a where

Class for values supporting an inner product and various norms.

Minimal complete definition

Nothing

Methods

(<.>) :: a -> a -> c infixl 7

Sum of component-wise product

norm_inf :: a -> c

Maximum absolute component

norm_0 :: a -> c

Number of non-zero components

norm_1 :: a -> c

Sum of absolute components

norm_2 :: a -> c

Square root of sum of squared components

quadrance :: a -> c

Sum of squared components

(<.>) :: (ADT a, Constraints a (Metric c)) => a -> a -> c infixl 7

Sum of component-wise product

norm_inf :: (ADT a, Constraints a (Metric c), Ord c) => a -> c

Maximum absolute component

norm_0 :: (ADT a, Constraints a (Metric c)) => a -> c

Number of non-zero components

norm_1 :: (ADT a, Constraints a (Metric c)) => a -> c

Sum of absolute components

norm_2 :: Floating c => a -> c

Square root of sum of squared components

quadrance :: (ADT a, Constraints a (Metric c)) => a -> c

Sum of squared components

gDot :: forall c a. (ADT a, Constraints a (Metric c), Num c) => a -> a -> c

An implementation of gDot that works for records where every field is an instance of Metric c.

gNorm_inf :: forall c a. (ADT a, Constraints a (Metric c), Ord c) => a -> c

An implementation of norm_inf that works for records where every field is an instance of Metric c.

gNorm_0 :: forall c a. (ADT a, Constraints a (Metric c), Num c) => a -> c

An implementation of norm_0 that works for records where every field is an instance of Metric c.

gNorm_1 :: forall c a. (ADT a, Constraints a (Metric c), Num c) => a -> c

An implementation of norm_1 that works for records where every field is an instance of Metric c.

gNorm_2 :: forall c a. (ADT a, Constraints a (Metric c), Floating c) => a -> c

An implementation of norm_2 that works for records where every field is an instance of Metric c.

gQuadrance :: forall c a. (ADT a, Constraints a (Metric c), Num c) => a -> c

An implementation of quadrance that works for records where every field is an instance of Metric c.

class (Mutable m a, Linear c a) => LinearInPlace m c a where

Instaces of Linear that support certain in-place mutations. Inspired by the BLAS Level 1 API. A LinearInPlace m v c a means that v is a mutable reference to an a that can be updated as an action in monad m.

Minimal complete definition

Nothing

Methods

(.+.=) :: Ref m a -> a -> m () infix 4

Add a value in-place.

(.*=) :: Ref m a -> c -> m () infix 4

Scale a value in-place.

(.*+=) :: Ref m a -> (c, a) -> m () infix 4

Add a scaled value in-place.

sumLinearInPlace :: (LinearInPlace m c a, Foldable t) => Ref m a -> t a -> m ()

Given some starting reference v, add every item in a foldable container into that reference in-place.

linearWit :: forall a c d. (Linear c a, Linear d a) => c :~: d

If a and b are both Linear instances, then if a is equal to b, their scalars c and d must also be equal. This is necessary because GHC isn't happy with the functional dependency for some reason.