Control.Monad.Sample

Typeclasses

class MonadPlus m => MonadSample r m | m -> r where

MonadPlus to imply that an empty pool is empty forever unless you re-fill it first.

Minimal complete definition

sample

Methods

sample :: m r

sampleN :: Vector v r => Int -> m (v r)

Should never "fail"

flushSamples :: MonadSample r m => m [r]

sampleReservoir :: forall m v r. (PrimMonad m, MonadSample r m, Vector v r) => Int -> Gen (PrimState m) -> m (v r)

Pulls samples until it cannot pull any more, and returns N shuffled items from the pool. Is O(N) memory.

Instances

newtype ConduitSample i o m a

Constructors

ConduitSample

Fields

runConduitSample :: ConduitT i o m (Maybe a)

conduitSample :: ConduitT i o m a -> ConduitSample i o m a

newtype RefSample v r m a

Constructors

RefSample

Fields

runRefSample :: v -> m (Maybe a)

newtype FoldSampleT rs m a

Constructors

FoldSampleT

Fields

runFoldSampleT :: rs -> m (Maybe a, rs)

Bundled Patterns

pattern FoldSample :: (rs -> (Maybe a, rs)) -> FoldSample rs a

type FoldSample rs = FoldSampleT rs Identity

foldSample :: Monad m => (rs -> (Maybe a, rs)) -> FoldSampleT rs m a