Numeric.Opto.Ref

Description

Abstract over different types for mutable references of values.

Documentation

class Monad m => Mutable m a where

Minimal complete definition

Nothing

Associated Types

type Ref m a = (v :: Type) | v -> a

Methods

thawRef :: a -> m (Ref m a)

freezeRef :: Ref m a -> m a

copyRef :: Ref m a -> a -> m ()

modifyRef :: Ref m a -> (a -> a) -> m ()

Apply a pure function on an immutable value onto a value stored in a mutable reference.

modifyRef' :: Ref m a -> (a -> a) -> m ()

modifyRef, but forces the result before storing it back in the reference.

updateRef :: Ref m a -> (a -> (a, b)) -> m b

Apply a pure function on an immutable value onto a value stored in a mutable reference, returning a result value from that function.

updateRef' :: Ref m a -> (a -> (a, b)) -> m b

updateRef, but forces the updated value before storing it back in the reference.

thawRef :: (Ref m a ~ MutVar (PrimState m) a, PrimMonad m) => a -> m (Ref m a)

freezeRef :: (Ref m a ~ MutVar (PrimState m) a, PrimMonad m) => Ref m a -> m a

copyRef :: (Ref m a ~ MutVar (PrimState m) a, PrimMonad m) => Ref m a -> a -> m ()

newtype MutRef a

Newtype wrapper that can provide any type with a Mutable instance. Can be useful for avoiding orphan instances.

Constructors

MutRef

Fields

runMutRef :: a

class Monad m => GMutable m f

Minimal complete definition

gThawRef_, gFreezeRef_, gCopyRef_

newtype GRef m a

Constructors

GRef

Fields

unGRef :: GRef_ m (Rep a) ()

gThawRef :: (Generic a, GMutable m (Rep a)) => a -> m (GRef m a)

gFreezeRef :: (Generic a, GMutable m (Rep a)) => GRef m a -> m a

gCopyRef :: (Generic a, GMutable m (Rep a)) => GRef m a -> a -> m ()

ReMutable

reMutable :: forall m n a r. (Mutable m a, Monad n) => (forall x. m x -> n x) -> (Mutable n a => r) -> r

If you can provice a natural transformation from m to n, you should be able to use a value as if it had Mutable n a if you have Mutable m a.

reMutableConstraint :: forall m n a. (Mutable m a, Monad n) => (forall x. m x -> n x) -> Mutable m a :- Mutable n a

If you can provice a natural transformation from m to n, then Mutable m a should also imply Mutable n a.

newtype ReMutable (s :: Type) m a

Constructors

ReMutable a

newtype ReMutableTrans m n

Constructors

RMT

Fields

runRMT :: forall x. m x -> n x