Numeric.Opto.Run

Description

Functions to run optimiziers.

Options

data RunOpts m a

Options for running an optimizer.

Constructors

RO

Fields

roStopCond :: Diff a -> a -> m Bool

roLimit :: Maybe Int

roBatch :: Maybe Int

noStop :: Applicative m => Maybe Int -> Maybe Int -> RunOpts m a

Construct a RunOpts without a stopping condition.

data ParallelOpts

Options for running an optimizer in a concurrent setting.

Constructors

PO

Fields

poThreads :: Maybe Int

poSplitRuns :: Int

how much each thread will process before regrouping

Single-threaded

runOpto :: forall m v a. Monad m => RunOpts m a -> a -> OptoM m v a -> m (a, OptoM m v a)

Optimize an a many times, until the stopping condition or limit is reached. Returns a frozen updated optimizer state.

evalOpto :: forall m v a. Monad m => RunOpts m a -> a -> OptoM m v a -> m a

Optimize an a many times, until the stopping condition or limit is reached. Returns a frozen updated optimizer state.

runOptoAlt :: forall m v a. MonadPlus m => RunOpts m a -> a -> OptoM m v a -> m (a, OptoM m v a)

A version of runOpto, but allowing the stopping condition to take advantage of an underlying Alternative instance, where 'empty'/'mzero' will quit immediately.

evalOptoAlt :: forall m v a. MonadPlus m => RunOpts m a -> a -> OptoM m v a -> m a

A version of evalOpto, but allowing the stopping condition to take advantage of an underlying Alternative instance, where 'empty'/'mzero' will quit immediately.

Parallel

runOptoParallel :: (MonadUnliftIO m, Fractional a) => RunOpts m a -> ParallelOpts -> a -> OptoM m v a -> m (a, [OptoM m v a])

Optimize an a times in parallel, and aggregate all the results together by taking the mean. Freezes and returns all of the updated optimizer states from each thread.

evalOptoParallel :: (MonadUnliftIO m, Fractional a) => RunOpts m a -> ParallelOpts -> a -> OptoM m v a -> m a