Numeric.Opto.Run

Description

Functions to run optimiziers.

Options

data RunOpts m a

Options for running an optimizer.

Constructors

RO

Fields

roStopCond :: Diff a -> a -> m Bool

roReport :: a -> m ()

reporting function

roLimit :: Maybe Int

number of batches to run (default = Nothing = until failure)

roBatch :: Maybe Int

batching updates (default = Nothing = no batching)

roFreq :: Maybe Int

batches per report (default = Nothing = every batch)

data ParallelOpts

Options for running an optimizer in a concurrent setting.

Constructors

PO

Fields

poThreads :: Maybe Int

Number of threads. Will default to max capacity

poSplitRuns :: Int

How much each thread will process before regrouping

Sampling

runOptoSample :: MonadSample r m => RunOpts m a -> a -> OptoM m v a -> m (a, OptoM m v a)

evalOptoSample :: MonadSample r m => RunOpts m a -> a -> OptoM m v a -> m a

Specific Sampling Monads

optoConduit :: Monad m => RunOpts (ConduitSample i a m) a -> a -> OptoM (ConduitSample i a m) v a -> ConduitT i a m (OptoM (ConduitSample i a m) v a)

optoConduit_ :: Monad m => RunOpts (ConduitSample i a m) a -> a -> OptoM (ConduitSample i a m) v a -> ConduitT i a m ()

refOpto :: (Monad m, Ref m rs vrs, Element rs ~ r, IsSequence rs, Index rs ~ Int) => RunOpts (RefSample vrs r m) a -> a -> OptoM (RefSample vrs r m) va a -> vrs -> m (a, OptoM (RefSample vrs r m) va a)

runOptoSample specialized for RefSample: give it a mutable reference vrs to a collection of items rs, and it will process each item r while emptying out the mutable reference.

It will terminate and return a result as soon as the reference is empty. You can add more items to the mutable reference as it is being processed, and it will process it as long as it hasn't yet completed.

Returns the optimized a, and a closure OptoM that can be resumed.

refOpto_ :: (Monad m, Ref m rs vrs, Element rs ~ r, IsSequence rs, Index rs ~ Int) => RunOpts (RefSample vrs r m) a -> a -> OptoM (RefSample vrs r m) va a -> vrs -> m a

foldOpto :: (Monad m, IsSequence rs, Index rs ~ Int) => RunOpts (FoldSampleT rs m) a -> a -> OptoM (FoldSampleT rs m) v a -> rs -> m (a, rs, OptoM (FoldSampleT rs m) v a)

runOptoSample specialized for FoldSampleT: give it a collection of items rs, and it will process each item r. Returns the optimized a, the leftover rs, and a closure OptoM that can be resumed.

foldOpto_ :: (Monad m, IsSequence rs, Index rs ~ Int) => RunOpts (FoldSampleT rs m) a -> a -> OptoM (FoldSampleT rs m) v a -> rs -> m (a, rs)

evalOptoSample specialized for FoldSampleT. Basically foldOpto, without returning the resumable closure.

Non-Sampling

runOpto :: Monad m => RunOpts m a -> a -> OptoM m v a -> m (a, OptoM m v a)

evalOpto :: Monad m => RunOpts m a -> a -> OptoM m v a -> m a