Numeric.Opto.Run

Description

Functions to run optimiziers.

Options

data RunOpts m a

Options for running an optimizer.

Constructors

RO

Fields

roStopCond :: Diff a -> a -> m Bool

Stop condition; will stop when True (default = never stop)

roReport :: a -> m ()

Reporting function (default = no report)

roLimit :: Maybe Int

number of batches to run (Nothing = run forever) (default = Nothing) | Size of batching updates (1 = no batching) (default = 1)

roBatch :: Int

roFreq :: Maybe Int

batches per report (Nothing = never report) (default = Just 1).

hoistRunOpts :: (forall x. m x -> n x) -> RunOpts m a -> RunOpts n a

data ParallelOpts

Options for running an optimizer in a concurrent setting.

Constructors

PO

Fields

poThreads :: Maybe Int

Number of threads (Nothing = max capacity) (default = Nothing)

poSplit :: Int

Single-threaded

runOpto :: Monad m => RunOpts m a -> m (Maybe r) -> a -> Opto m v r a -> m (a, Opto m v r a)

evalOpto :: Monad m => RunOpts m a -> m (Maybe r) -> a -> Opto m v r a -> m a

runOptoNonSampling :: Monad m => RunOpts m a -> a -> Opto m v () a -> m (a, Opto m v () a)

evalOptoNonSampling :: Monad m => RunOpts m a -> a -> Opto m v () a -> m a

Sampling methods

optoConduit :: Monad m => RunOpts m a -> a -> Opto (ConduitT r a m) v r a -> ConduitT r a m (Opto (ConduitT r a m) v r a)

optoConduit_ :: Monad m => RunOpts m a -> a -> Opto (ConduitT r a m) v r a -> ConduitT r a m ()

foldOpto :: (Monad m, IsSequence rs, r ~ Element rs) => RunOpts m a -> a -> Opto (StateT rs m) v r a -> rs -> m (a, rs, Opto (StateT rs m) v r a)

runOptoSample specialized for FoldSampleT: give it a collection of items rs, and it will process each item r. Returns the optimized a, the leftover rs, and a closure Opto that can be resumed.

foldOpto_ :: (Monad m, IsSequence rs, r ~ Element rs) => RunOpts m a -> a -> Opto (StateT rs m) v r a -> rs -> m (a, rs)

Parallel

evalOptoParallel :: forall m v r a. (MonadUnliftIO m, Fractional a) => RunOpts m a -> ParallelOpts -> m (Maybe r) -> a -> Opto m v r a -> m a

evalOptoParallelChunk :: forall m v r a rs. (MonadUnliftIO m, Fractional a, IsSequence rs, r ~ Element rs) => RunOpts m a -> ParallelOpts -> (Int -> m rs) -> a -> Opto (StateT rs m) v r a -> m a

Sampling Methods

optoConduitParallel :: forall m v r a. (MonadUnliftIO m, Fractional a) => RunOpts m a -> ParallelOpts -> a -> Opto m v r a -> ConduitT () r m () -> ConduitT () a m ()

optoConduitParallelChunk :: forall m v r a. (MonadUnliftIO m, Fractional a) => RunOpts m a -> ParallelOpts -> a -> Opto (StateT [r] m) v r a -> ConduitT () r m () -> ConduitT () a m ()