Numeric.Opto.Run

Description

Functions to run optimiziers.

Options

data RunOpts m a

Options for running an optimizer.

Constructors

RO

Fields

roStopCond :: Diff a -> a -> m Bool

Stop condition; will stop when True (default = never stop)

roReport :: a -> m ()

Reporting function (default = no report)

roLimit :: Maybe Int

Number of batches to run (Nothing = run forever) (default = Nothing).

roBatch :: Int

Size of batching updates (1 = no batching) (default = 1)

roFreq :: Maybe Int

batches per report (Nothing = never report) (default = Just 1).

hoistRunOpts :: (forall x. m x -> n x) -> RunOpts m a -> RunOpts n a

data ParallelOpts m a

Options for running an optimizer in a concurrent setting.

Constructors

PO

Fields

poThreads :: Maybe Int

Number of threads (Nothing = max capacity) (default = Nothing)

poSplit :: Int

How to recombine a pool of updated results into a single result (default = pure . mean)

poCombine :: NonEmpty a -> m a

hoistParallelOpts :: (forall x. m x -> n x) -> ParallelOpts m a -> ParallelOpts n a

Single-threaded

opto :: Monad m => RunOpts m a -> m (Maybe r) -> a -> Opto m v r a -> m a

opto' :: Monad m => RunOpts m a -> m (Maybe r) -> a -> Opto m v r a -> m (a, Opto m v r a)

optoNonSampling' :: Monad m => RunOpts m a -> a -> Opto m v () a -> m (a, Opto m v () a)

optoNonSampling :: Monad m => RunOpts m a -> a -> Opto m v () a -> m a

Sampling methods

optoConduit :: Monad m => RunOpts m a -> a -> Opto (ConduitT r a m) v r a -> ConduitT r a m ()

optoConduit' :: Monad m => RunOpts m a -> a -> Opto (ConduitT r a m) v r a -> ConduitT r a m (Opto (ConduitT r a m) v r a)

optoFold :: (Monad m, IsSequence rs, r ~ Element rs) => RunOpts m a -> a -> Opto (StateT rs m) v r a -> rs -> m (a, rs)

optoFold' :: (Monad m, IsSequence rs, r ~ Element rs) => RunOpts m a -> a -> Opto (StateT rs m) v r a -> rs -> m (a, rs, Opto (StateT rs m) v r a)

runOptoSample specialized for FoldSampleT: give it a collection of items rs, and it will process each item r. Returns the optimized a, the leftover rs, and a closure Opto that can be resumed.

Parallel

optoPar :: forall m v r a. MonadUnliftIO m => RunOpts m a -> ParallelOpts m a -> m (Maybe r) -> a -> Opto m v r a -> m a

optoParChunk :: forall m v r a rs. (MonadUnliftIO m, IsSequence rs, r ~ Element rs) => RunOpts m a -> ParallelOpts m a -> (Int -> m rs) -> a -> Opto (StateT rs m) v r a -> m a

optoParNonSampling :: MonadUnliftIO m => RunOpts m a -> ParallelOpts m a -> a -> Opto m v () a -> m a

Sampling Methods

optoConduitPar :: forall m v r a. MonadUnliftIO m => RunOpts m a -> ParallelOpts m a -> a -> Opto m v r a -> ConduitT () r m () -> ConduitT () a m ()

optoConduitParChunk :: forall m v r a. MonadUnliftIO m => RunOpts m a -> ParallelOpts m a -> a -> Opto (StateT [r] m) v r a -> ConduitT () r m () -> ConduitT () a m ()

Util

mean :: (Foldable1 t, Fractional a) => t a -> a