Numeric.Opto.Run.Conduit

Running Optos

Single threaded

data RunOpts m a

Options for running an optimizer.

Constructors

RO

Fields

roStopCond :: Diff a -> a -> m Bool

roLimit :: Maybe Int

roBatch :: Maybe Int

runOptoConduit :: Monad m => RunOpts (SampleT r (ConduitT r a m)) a -> a -> OptoM (SampleT r (ConduitT r a m)) v a -> ConduitT r a m (OptoM (SampleT r (ConduitT r a m)) v a)

With RunOpts, a chunk size, an initial input, and a sampling optimizer, give a conduit that processes upstream samples and outputs every single one of the updated values as they are generated.

Returns the updated optimizer state.

runOptoConduit_ :: Monad m => RunOpts (SampleT r (ConduitT r a m)) a -> a -> OptoM (SampleT r (ConduitT r a m)) v a -> ConduitT r a m ()

runOptoConduit, without returning the updated optimizer state.

runOptoConduitChunk :: Monad m => RunOpts (SampleT r (ConduitT r a m)) a -> a -> OptoM (SampleT r (ConduitT r a m)) v a -> ConduitT r a m (OptoM (SampleT r (ConduitT r a m)) v a)

With RunOpts, a chunk size, an initial input, and a sampling optimizer, give a conduit that processes upstream samples and outputs the updated value only after the optimizer finishes.

Returns the updated optimizer state.

runOptoConduitChunk_ :: Monad m => RunOpts (SampleT r (ConduitT r a m)) a -> a -> OptoM (SampleT r (ConduitT r a m)) v a -> ConduitT r a m ()

runOptoConduitChunk, without returning the updated optimizer state.

Parallel

data ParallelOpts

Options for running an optimizer in a concurrent setting.

Constructors

PO

Fields

poThreads :: Maybe Int

poSplitRuns :: Int

how much each thread will process before regrouping

Sampling conduits

shuffling :: PrimMonad m => Gen (PrimState m) -> ConduitT a a m ()

Outputs a shuffled version of the input stream. Keeps entire input stream in memory.

NOTE: Pulls the entire input stream into memory first before outputting anything.

shufflingN :: PrimMonad m => Int -> Gen (PrimState m) -> ConduitT a a m ()

Takes the first N items out of the input stream, shuffles them in-memory, and outputs the shuffled result.

Leaves the rest of the items in the stream.

Use forever to repeat multiple times until the stream is exhausted.

sinkSampleReservoir :: forall m v a o. (PrimMonad m, Vector v a) => Int -> Gen (PrimState m) -> ConduitT a o m (v a)

Process an entire stream, and keep N random and shuffled items from that stream. Is O(N) memory.

samplingN :: PrimMonad m => Int -> Gen (PrimState m) -> ConduitT a a m ()

Process an entire stream, and yield N random items from that stream. Is O(N) memory.

NOTE: Exhausts the entire input stream first before outputting anything, but never keeps the entire stream in memory.

skipSampling

Arguments

:: PrimMonad m

=> Double
probability of pass-through

-> Gen (PrimState m)

-> ConduitT a a m ()

Drops and lets items through randomly with a given probability.