Numeric.Opto.Run.Conduit

Documentation

runOptoConduit :: Monad m => RunOpts (SampleConduit i a m) a -> a -> OptoM (SampleConduit i a m) v a -> ConduitT i a m (OptoM (SampleConduit i a m) v a)

shuffling :: PrimMonad m => Gen (PrimState m) -> ConduitT a a m ()

Outputs a shuffled version of the input stream. Keeps entire input stream in memory.

NOTE: Pulls the input stream first before outputting anything.

shufflingN :: PrimMonad m => Int -> Gen (PrimState m) -> ConduitT a a m ()

Takes the first N items out of the input stream, shuffles them in-memory, and outputs the shuffled result.

Leaves the rest of the items in the stream.

Use forever to repeat multiple times until the stream is exhausted.

sinkSampleReservoir :: forall m v a o. (PrimMonad m, Vector v a) => Int -> Gen (PrimState m) -> ConduitT a o m (v a)

samplingN :: PrimMonad m => Int -> Gen (PrimState m) -> ConduitT a a m ()